requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asn1crypto==1.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
starlette==0.41.3
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
python-multipart